refactor: change tokens type to entity, replace responses with statuses with new response decorators
