+++ src/auth/auth.controller.ts
@@ -9,9 +9,16 @@ import {
   Get,
 } from '@nestjs/common';
 import { Response } from 'express';
-import { ApiOperation, ApiResponse, ApiTags } from '@nestjs/swagger';
+import {
+  ApiBadRequestResponse,
+  ApiCreatedResponse,
+  ApiOkResponse,
+  ApiOperation,
+  ApiTags,
+  ApiUnauthorizedResponse,
+} from '@nestjs/swagger';
 
-import { Tokens } from './types/tokens.type';
+import { Tokens } from './entities/tokens.entity';
 import { AuthService } from './auth.service';
 import { AuthDto } from './dto/auth.dto';
 import { GetCurrentUser } from 'src/common/decorators/get-current-user.decorator';
@@ -26,11 +33,13 @@ export class AuthController {
   @Post('signup')
   @HttpCode(HttpStatus.CREATED)
   @ApiOperation({ summary: 'Creates the user and returns tokens' })
-  @ApiResponse({
-    status: HttpStatus.CREATED,
+  @ApiCreatedResponse({
     description: 'User created',
+    type: Tokens,
   })
-  @ApiResponse({ status: HttpStatus.BAD_REQUEST, description: 'Bad Request' })
+  @ApiBadRequestResponse({
+    description: 'Bad Request',
+  })
   signup(
     @Body() dto: AuthDto,
     @Res({ passthrough: true }) res: Response,
@@ -41,9 +50,8 @@ export class AuthController {
   @Post('signin')
   @HttpCode(HttpStatus.OK)
   @ApiOperation({ summary: 'Returns tokens if user exists' })
-  @ApiResponse({ status: HttpStatus.OK, description: 'User exists' })
-  @ApiResponse({
-    status: HttpStatus.UNAUTHORIZED,
+  @ApiOkResponse({ description: 'User exists', type: Tokens })
+  @ApiUnauthorizedResponse({
     description: 'User not exists',
   })
   signin(
@@ -57,12 +65,10 @@ export class AuthController {
   @UseGuards(CustomAuthGuard)
   @HttpCode(HttpStatus.OK)
   @ApiOperation({ summary: 'Clears user tokens' })
-  @ApiResponse({
-    status: HttpStatus.OK,
+  @ApiOkResponse({
     description: 'User tokens are cleared',
   })
-  @ApiResponse({
-    status: HttpStatus.UNAUTHORIZED,
+  @ApiUnauthorizedResponse({
     description: 'User is not authorized',
   })
   logout(@Res({ passthrough: true }) res: Response) {
@@ -73,13 +79,11 @@ export class AuthController {
   @UseGuards(CustomAuthGuard)
   @HttpCode(HttpStatus.OK)
   @ApiOperation({ summary: 'Returns user profile' })
-  @ApiResponse({
-    status: HttpStatus.OK,
+  @ApiOkResponse({
     description: 'User profile returned',
     type: User,
   })
-  @ApiResponse({
-    status: HttpStatus.UNAUTHORIZED,
+  @ApiUnauthorizedResponse({
     description: 'User is not authorized',
   })
   getProfile(@GetCurrentUser('sub') userId: number): User {
