+++ src/auth/dto/auth.dto.ts
@@ -4,6 +4,7 @@ import { IsEmail, IsNotEmpty, IsString } from 'class-validator';
 export class AuthDto {
   @ApiProperty({
     name: 'email',
+    description: 'email of the user',
     minLength: 1,
     example: 'example@mail.com',
   })
@@ -13,6 +14,7 @@ export class AuthDto {
 
   @ApiProperty({
     name: 'password',
+    description: 'password of the user',
     minLength: 1,
     example: '1234567',
   })
