+++ src/auth/entities/tokens.entity.ts
@@ -1,0 +1,13 @@
+import { ApiProperty } from '@nestjs/swagger';
+
+export class Tokens {
+  @ApiProperty({
+    description: 'access token',
+  })
+  accessToken: string;
+
+  @ApiProperty({
+    description: 'refresh token',
+  })
+  refreshToken: string;
+}
