--- src/auth/types/tokens.type.ts
@@ -1,4 +1,0 @@
-export type Tokens = {
-  accessToken: string;
-  refreshToken: string;
-};
